requirements.txt
Flask<1
azure==0.8.4
mysql-connector==2.1.4